+++ integration/logtalk_comp_xsb.pl
@@ -41,5 +41,5 @@
 #include ../core/core.pl
 
 % workaround the lack of support for static multifile predicates
-:- dynamic('$lgt_logtalk.debug_handler_provider'/2).
-:- dynamic('$lgt_logtalk.debug_handler'/3).
+:- dynamic('$logtalk.debug_handler_provider'/2).
+:- dynamic('$logtalk.debug_handler'/3).